Move logger into cli project, allow caller to close underlying filehandle
