+++ src/packages/cli/src/cli.ts
@@ -6,6 +6,7 @@ import args from "./args";
 import { EthereumFlavorName, FilecoinFlavorName } from "@ganache/flavors";
 import initializeEthereum from "./initialize/ethereum";
 import initializeFilecoin from "./initialize/filecoin";
+import { createLogger } from "./logger";
 import type { FilecoinProvider } from "@ganache/filecoin";
 import type { EthereumProvider } from "@ganache/ethereum";
 import {
@@ -54,6 +55,14 @@ if (argv.action === "start") {
   const flavor = argv.flavor;
   const cliSettings = argv.server;
 
+  const loggingOptions = (argv as any).logging;
+  let closeLogHandle = null;
+  if (loggingOptions.file) {
+    const { log, close } = createLogger((argv as any).logging);
+    (argv as any).logging.logger = { log };
+    closeLogHandle = close;
+  }
+
   console.log(detailedVersion);
 
   let server: ReturnType<typeof Ganache.server>;
@@ -88,7 +97,12 @@ if (argv.action === "start") {
           return;
         case ServerStatus.open:
           console.log("Shutting down…");
-          await server.close();
+
+          if (closeLogHandle) {
+            await Promise.all([server.close(), closeLogHandle()]);
+          } else {
+            await server.close();
+          }
           console.log("Server has been shut down");
           break;
       }
